+++ jenkins-client/src/main/java/com/offbytwo/jenkins/client/JenkinsHttpClient.java
@@ -3,7 +3,6 @@
  *
  * Distributed under the MIT license: http://opensource.org/licenses/MIT
  */
-
 package com.offbytwo.jenkins.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -57,9 +56,8 @@ import com.offbytwo.jenkins.client.util.ResponseUtils;
 import com.offbytwo.jenkins.client.util.UrlUtils;
 import static org.apache.commons.lang.StringUtils.isNotBlank;
 
-//import com.offbytwo.jenkins.client.util.HttpResponseContentExtractor;
+public class JenkinsHttpClient implements JenkinsHttpConnection {
 
-public class JenkinsHttpClient {
     private final Logger LOGGER = LoggerFactory.getLogger(getClass());
 
     private URI uri;
@@ -74,7 +72,7 @@ public class JenkinsHttpClient {
     private String jenkinsVersion;
 
     public final static String EMPTY_VERSION = "UNKNOWN";
-    
+
     /**
      * Create an unauthenticated Jenkins HTTP client
      *
@@ -142,14 +140,9 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a GET request and parse the response to the given class
-     *
-     * @param path path to request, can be relative or absolute
-     * @param cls class of the response
-     * @param <T> type of the response
-     * @return an instance of the supplied class
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public <T extends BaseModel> T get(String path, Class<T> cls) throws IOException {
         HttpGet getMethod = new HttpGet(UrlUtils.toJsonApiUri(uri, context, path));
 
@@ -165,13 +158,9 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a GET request and parse the response and return a simple string
-     * of the content
-     *
-     * @param path path to request, can be relative or absolute
-     * @return the entity text
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String get(String path) throws IOException {
         HttpGet getMethod = new HttpGet(UrlUtils.toJsonApiUri(uri, context, path));
         HttpResponse response = client.execute(getMethod, localContext);
@@ -189,14 +178,9 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a GET request and parse the response to the given class, logging
-     * any IOException that is thrown rather than propagating it.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param cls class of the response
-     * @param <T> type of the response
-     * @return an instance of the supplied class
+     * {@inheritDoc}
      */
+    @Override
     public <T extends BaseModel> T getQuietly(String path, Class<T> cls) {
         T value;
         try {
@@ -210,12 +194,9 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a GET request and return the response as InputStream
-     *
-     * @param path path to request, can be relative or absolute
-     * @return the response stream
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public InputStream getFile(URI path) throws IOException {
         HttpGet getMethod = new HttpGet(path);
         HttpResponse response = client.execute(getMethod, localContext);
@@ -224,23 +205,19 @@ public class JenkinsHttpClient {
         return new RequestReleasingInputStream(response.getEntity().getContent(), getMethod);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public <R extends BaseModel, D> R post(String path, D data, Class<R> cls) throws IOException {
         return post(path, data, cls, null, true);
     }
 
     /**
-     * Perform a POST request and parse the response to the given class
-     *
-     * @param path path to request, can be relative or absolute
-     * @param data data to post
-     * @param cls class of the response
-     * @param <R> type of the response
-     * @param <D> type of the data
+     * {@inheritDoc}
      * @param fileParams the job file parameters
-     * @param crumbFlag true / false.
-     * @return an instance of the supplied class
-     * @throws IOException in case of an error.
      */
+    @Override
     public <R extends BaseModel, D> R post(String path, D data, Class<R> cls, Map<String, File> fileParams, boolean crumbFlag) throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
         if (crumbFlag == true) {
@@ -296,23 +273,9 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a POST request using form url encoding.
-     * 
-     * This method was added for the purposes of creating folders, but may be
- useful for other API calls as well.
- 
- Unlike post and post_xml, the path is *not* modified by adding
- "/toJsonApiUri/json". Additionally, the params in data are provided as both
- request parameters including a json parameter, *and* in the
- JSON-formatted StringEntity, because this is what the folder creation
- call required. It is unclear if any other jenkins APIs operate in this
- fashion.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param data data to post
-     * @param crumbFlag true / false.
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public void post_form(String path, Map<String, String> data, boolean crumbFlag) throws IOException {
         HttpPost request;
         if (data != null) {
@@ -350,18 +313,10 @@ public class JenkinsHttpClient {
         }
     }
 
-
     /**
-     * Perform a POST request using form url encoding and return HttpResponse object
-     * This method is not performing validation and can be used for more generic queries to jenkins.
-     *
-     * @param path
-     *            path to request, can be relative or absolute
-     * @param data
-     *            data to post
-     * @throws IOException,
-     *             HttpResponseException
+     * {@inheritDoc}
      */
+    @Override
     public HttpResponse post_form_with_result(String path, List<NameValuePair> data, boolean crumbFlag) throws IOException {
         HttpPost request;
         if (data != null) {
@@ -384,18 +339,17 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform a POST request of XML (instead of using json mapper) and return a
-     * string rendering of the response entity.
-     *
-     * @param path path to request, can be relative or absolute
-     * @param xml_data data data to post
-     * @return A string containing the xml response (if present)
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_xml(String path, String xml_data) throws IOException {
         return post_xml(path, xml_data, true);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public String post_xml(String path, String xml_data, boolean crumbFlag) throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
         if (crumbFlag == true) {
@@ -420,28 +374,17 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Post a text entity to the given URL using the default content type
-     *
-     * @param path The path.
-     * @param textData data.
-     * @param crumbFlag true/false.
-     * @return resulting response
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_text(String path, String textData, boolean crumbFlag) throws IOException {
         return post_text(path, textData, ContentType.DEFAULT_TEXT, crumbFlag);
     }
 
     /**
-     * Post a text entity to the given URL with the given content type
-     *
-     * @param path The path.
-     * @param textData The data.
-     * @param contentType {@link ContentType}
-     * @param crumbFlag true or false.
-     * @return resulting response
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public String post_text(String path, String textData, ContentType contentType, boolean crumbFlag)
             throws IOException {
         HttpPost request = new HttpPost(UrlUtils.toJsonApiUri(uri, context, path));
@@ -467,25 +410,97 @@ public class JenkinsHttpClient {
     }
 
     /**
-     * Perform POST request that takes no parameters and returns no response
-     *
-     * @param path path to request
-     * @throws IOException in case of an error.
+     * {@inheritDoc}
      */
+    @Override
     public void post(String path) throws IOException {
         post(path, null, null, null, false);
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    @Override
     public void post(String path, boolean crumbFlag) throws IOException {
         post(path, null, null,null, crumbFlag);
     }
 
-    
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String getJenkinsVersion() {
+        return this.jenkinsVersion;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean isJenkinsVersionSet() {
+        return !EMPTY_VERSION.equals(this.jenkinsVersion);
+    }
+
+    /**
+     * Closes underlying resources. Any I/O errors whilst closing are logged
+     * with debug log level Closed instances should no longer be used Closing an
+     * already closed instance has no side effects
+     */
+    @Override
+    public void close() {
+        try {
+            client.close();
+        } catch (final IOException ex) {
+            LOGGER.debug("I/O exception closing client", ex);
+        }
+    }
 
     
+    /**
+     * Add authentication to supplied builder.
+     * @param builder the builder to configure
+     * @param uri the server URI
+     * @param username the username
+     * @param password the password
+     * @return the passed in builder
+     */
+    protected static HttpClientBuilder addAuthentication(final HttpClientBuilder builder, 
+            final URI uri, final String username,
+            String password) {
+        if (isNotBlank(username)) {
+            CredentialsProvider provider = new BasicCredentialsProvider();
+            AuthScope scope = new AuthScope(uri.getHost(), uri.getPort(), "realm");
+            UsernamePasswordCredentials credentials = new UsernamePasswordCredentials(username, password);
+            provider.setCredentials(scope, credentials);
+            builder.setDefaultCredentialsProvider(provider);
+
+            builder.addInterceptorFirst(new PreemptiveAuth());
+        }
+        return builder;
+    }
 
     
+    /**
+     * Get the local context.
+     * @return context
+     */
+    protected HttpContext getLocalContext() {
+        return localContext;
+    }
 
+    
+    /**
+     * Set the local context.
+     * @param localContext the context
+     */
+    protected void setLocalContext(final HttpContext localContext) {
+        this.localContext = localContext;
+    }
+
+    
+    
+    
+    
     private <T extends BaseModel> T objectFromResponse(Class<T> cls, HttpResponse response) throws IOException {
         InputStream content = response.getEntity().getContent();
         byte[] bytes = ByteStreams.toByteArray(content);
@@ -502,48 +517,8 @@ public class JenkinsHttpClient {
         return mapper;
     }
 
-    /**
-     * @return the version string.
-     */
-    public String getJenkinsVersion() {
-        return this.jenkinsVersion;
-    }
-
-    /**
-     * Check to see if the jenkins version has been set
-     * to something different than the initialization value
-     * from the constructor. This means there has never been made
-     * a communication with the Jenkins server.
-     * @return true if jenkinsVersion has been set by communication, false otherwise. 
-     */
-    public boolean isJenkinsVersionSet() {
-        return !EMPTY_VERSION.equals( this.jenkinsVersion );
-    }
-    
-
     private void releaseConnection(HttpRequestBase httpRequestBase) {
         httpRequestBase.releaseConnection();
     }
 
-    protected static HttpClientBuilder addAuthentication(HttpClientBuilder builder, URI uri, String username,
-            String password) {
-        if (isNotBlank(username)) {
-            CredentialsProvider provider = new BasicCredentialsProvider();
-            AuthScope scope = new AuthScope(uri.getHost(), uri.getPort(), "realm");
-            UsernamePasswordCredentials credentials = new UsernamePasswordCredentials(username, password);
-            provider.setCredentials(scope, credentials);
-            builder.setDefaultCredentialsProvider(provider);
-
-            builder.addInterceptorFirst(new PreemptiveAuth());
-        }
-        return builder;
-    }
-
-    protected HttpContext getLocalContext() {
-        return localContext;
-    }
-
-    protected void setLocalContext(HttpContext localContext) {
-        this.localContext = localContext;
-    }
 }