+++ ReleaseNotes.md
@@ -2,7 +2,25 @@
 
 ## Release 0.3.8 (NOT RELEASED YET)
 
- * [JENKINS-46472](https://issues.jenkins-ci.org/browse/JENKINS-46472)
+ * [Fixed Issue 289][issue-289]
+   
+   Added a build.stop() method which takes in a crumbFlag
+
+
+ * [Fixed Issue 301][issue-301]
+   
+   Decoupled JenkinsServer and JenkinsHttpClient by extracting JenkinsHttpClient 
+   methods into public interface so that different implementations can be plugged
+   into JenkinsServer if required
+
+
+ * [Fixed Issue 298][issue-298]
+   
+   Added Closeable support to JenkinsServer and JenkinsHttpClient so that
+   underlying resources can be cleaned up when instances no longer required
+
+
+ * [JENKINS-46472][jissue-46472]
 
    Added ability to modify offline cause for offline computers.
 
@@ -14,7 +32,7 @@
     }
    ```
 
- * [JENKINS-46445](https://issues.jenkins-ci.org/browse/JENKINS-46445)
+ * [JENKINS-46445][jissue-46445]
 
    Add support for both client TLS and basic authentication.
 
@@ -1043,6 +1061,8 @@ TestReport testReport = mavenJob.getLastSuccessfulBuild().getTestReport();
 [issue-222]: https://github.com/jenkinsci/java-client-api/issues/222
 [issue-244]: https://github.com/jenkinsci/java-client-api/issues/244
 [issue-268]: https://github.com/jenkinsci/java-client-api/issues/268
+[issue-298]: https://github.com/jenkinsci/java-client-api/issues/298
+[issue-301]: https://github.com/jenkinsci/java-client-api/issues/301
 [pull-123]: https://github.com/jenkinsci/java-client-api/pull/123
 [pull-149]: https://github.com/jenkinsci/java-client-api/pull/149
 [pull-158]: https://github.com/jenkinsci/java-client-api/pull/158
@@ -1059,3 +1079,5 @@ TestReport testReport = mavenJob.getLastSuccessfulBuild().getTestReport();
 [jissue-38787]: https://issues.jenkins-ci.org/browse/JENKINS-38787
 [jissue-38816]: https://issues.jenkins-ci.org/browse/JENKINS-38816
 [jissue-38823]: https://issues.jenkins-ci.org/browse/JENKINS-38823
+[jissue-46445]: https://issues.jenkins-ci.org/browse/JENKINS-46445)
+[jissue-46472]: https://issues.jenkins-ci.org/browse/JENKINS-46472)